webapp/requirements.txt
attrs==21.4.0
click==8.1.3
Flask==2.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MyApplication==0.1.0
packaging==21.3
pluggy==1.0.0
psycopg==3.0.15
psycopg-binary==3.0.15
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.8.2
PyYAML==6.0
tomli==2.0.1
Werkzeug==2.1.2
zipp==3.8.0
